requirements.txt
nomic
tweet-preprocessor
snscrape
tqdm
nltk
loguru
numpy
sklearn
jsonlines